pyproject.toml
[tool.poetry]
name = "python-drf-test"
version = "0.1.0"
description = ""
authors = ["DaniilGamers <ss.mihalkivd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
